services/ml_worker/requirements.txt
pika>=1.1.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.3
python-dotenv>=0.19.0
pandas>=1.3.5
numpy>=1.22.0
scikit-learn>=1.0.2
joblib>=1.1.0
bcrypt==4.0.1
deepface==0.0.79
pillow==9.5.0
opencv-python-headless==4.8.0.74
tensorflow==2.12.0 